+++ api/src/main/java/com/yanny/ali/api/ICommonRegistry.java
@@ -8,5 +8,5 @@ import java.util.List;
 import java.util.function.Function;
 
 public interface ICommonRegistry {
-    <T extends Entity> void registerEntityVariants(EntityType<?> type, Function<Level, List<Entity>> factory);
+    <T extends Entity> void registerEntityVariants(EntityType<T> type, Function<Level, List<Entity>> factory);
 }