+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## []
+
+- Fixed Ingredient tooltip
+- Make uppercased chest name from loot table name if missing translation
+
+## [1.2.3]
+
+- Fixed crash on failed mixin
+
 ## [1.2.2]
 
 - Fixed crash when empty itemStack